# Optimize the `images()` function: ~2 orders of magnitude faster unit tests