This commit makes two modifications to the `image()` function:
1) Read the whole MNIST data file in memory before working on the data;
2) Return the read data instead of unwrapping every elements and recreating
   a vector from it.

The inefficiency comes from the the `bytes()` method (from
the `std::io::Read` trait) returns an iterator over the bytes, but
wrapped in an `Option`. Calling it before map().collect() would read the
whole file, but would wrap every byte in an `Option` before unwrapping it
and reallocating a new vector, which is highly ineficient.

Instead, read the whole file in memory first as a single operation. The
result is a vector of bytes. A slice of it is taken to be used by the
`byteorder` crate's `ReadBytesExt` trait (`read_u32()`).

This gives a ~65 times speedup on the included unit tests when run in
release mode. Note that the unit tests can take more than 60 seconds in
debug mode and thus are not compared. Tests were run on a MacBook Air 2013.

Before:
```
 -> cargo build --release && export RUST_TEST_THREADS=1 && time cargo test --release
    Finished release [optimized] target(s) in 0.0 secs
    Finished release [optimized] target(s) in 0.0 secs
     Running /Users/nicolas/mnist/target/release/deps/mnist-ce337fc2c160ab7d

running 14 tests
test tests::test_00 ... ok
test tests::test_01 ... ok
test tests::test_02 ... ok
test tests::test_03 ... ok
test tests::test_04 ... ok
test tests::test_05 ... ok
test tests::test_06 ... ok
test tests::test_07 ... ok
test tests::test_08 ... ok
test tests::test_09 ... ok
test tests::test_10 ... ok
test tests::test_11 ... ok
test tests::test_12 ... ok
test tests::test_example ... ok

test result: ok. 14 passed; 0 failed; 0 ignored; 0 measured

   Doc-tests mnist

running 13 tests
test MnistBuilder<'a>::base_path_0 ... ok
test MnistBuilder<'a>::finalize_0 ... ok
test MnistBuilder<'a>::label_format_digit_0 ... ok
test MnistBuilder<'a>::label_format_one_hot_0 ... ok
test MnistBuilder<'a>::new_0 ... ok
test MnistBuilder<'a>::test_images_filename_0 ... ok
test MnistBuilder<'a>::test_labels_filename_0 ... ok
test MnistBuilder<'a>::test_set_length_0 ... ok
test MnistBuilder<'a>::training_images_filename_0 ... ok
test MnistBuilder<'a>::training_labels_filename_0 ... ok
test MnistBuilder<'a>::training_set_length_0 ... ok
test MnistBuilder<'a>::validation_set_length_0 ... ok
test _0 ... ok

test result: ok. 13 passed; 0 failed; 0 ignored; 0 measured

real    3m43.705s
user    0m43.770s
sys     2m54.431s
```

After:
```
 -> cargo build --release && export RUST_TEST_THREADS=1 && time cargo test --release
   Compiling mnist v0.2.0 (file:///Users/nicolas/mnist)
    Finished release [optimized] target(s) in 0.88 secs
   Compiling mnist v0.2.0 (file:///Users/nicolas/mnist)
    Finished release [optimized] target(s) in 1.41 secs
     Running /Users/nicolas/target/release/deps/mnist-ce337fc2c160ab7d

running 14 tests
test tests::test_00 ... ok
test tests::test_01 ... ok
test tests::test_02 ... ok
test tests::test_03 ... ok
test tests::test_04 ... ok
test tests::test_05 ... ok
test tests::test_06 ... ok
test tests::test_07 ... ok
test tests::test_08 ... ok
test tests::test_09 ... ok
test tests::test_10 ... ok
test tests::test_11 ... ok
test tests::test_12 ... ok
test tests::test_example ... ok

test result: ok. 14 passed; 0 failed; 0 ignored; 0 measured

   Doc-tests mnist

running 13 tests
test MnistBuilder<'a>::base_path_0 ... ok
test MnistBuilder<'a>::finalize_0 ... ok
test MnistBuilder<'a>::label_format_digit_0 ... ok
test MnistBuilder<'a>::label_format_one_hot_0 ... ok
test MnistBuilder<'a>::new_0 ... ok
test MnistBuilder<'a>::test_images_filename_0 ... ok
test MnistBuilder<'a>::test_labels_filename_0 ... ok
test MnistBuilder<'a>::test_set_length_0 ... ok
test MnistBuilder<'a>::training_images_filename_0 ... ok
test MnistBuilder<'a>::training_labels_filename_0 ... ok
test MnistBuilder<'a>::training_set_length_0 ... ok
test MnistBuilder<'a>::validation_set_length_0 ... ok
test _0 ... ok

test result: ok. 13 passed; 0 failed; 0 ignored; 0 measured

real    0m3.481s
user    0m2.139s
sys     0m1.201s
```